# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  article.CreateArticleCommand:
    properties:
      body:
        type: string
      categoryIds:
        items:
          type: integer
        type: array
      description:
        type: string
      mediaIds:
        items:
          type: integer
        type: array
      seoTags:
        items:
          type: string
        type: array
      siteId:
        type: integer
      slug:
        type: string
      title:
        type: string
    required:
    - siteId
    type: object
  article.DeleteArticleCommand:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  article.GetByFiltersSortArticleQuery:
    properties:
      page:
        minimum: 1
        type: integer
      pageSize:
        maximum: 100
        minimum: 1
        type: integer
      search:
        type: string
      searchBy:
        enum:
        - title
        - description
        type: string
      selectedFilters:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      selectedSort:
        $ref: '#/definitions/enums.ArticleSortEnum'
      siteId:
        type: integer
      sort:
        enum:
        - asc
        - desc
        type: string
      sortBy:
        enum:
        - title
        - description
        type: string
    required:
    - page
    - pageSize
    - siteId
    type: object
  article.UpdateArticleCommand:
    properties:
      body:
        type: string
      categoryIds:
        items:
          type: integer
        type: array
      description:
        type: string
      id:
        type: integer
      mediaIds:
        items:
          type: integer
        type: array
      seoTags:
        items:
          type: string
        type: array
      siteId:
        type: integer
      slug:
        type: string
      title:
        type: string
    required:
    - id
    - siteId
    type: object
...
  enums.ArticleSortEnum:
    enum:
    - title_a_z
    - title_z_a
    - recently_added
    - recently_updated
    - most_visited
    - least_visited
    - most_rated
    - least_rated
    - most_reviewed
    - least_reviewed
    type: string
    x-enum-varnames:
    - ArticleTitleAZSort
    - ArticleTitleZASort
    - ArticleRecentlyAddedSort
    - ArticleRecentlyUpdatedSort
    - ArticleMostVisitedSort
    - ArticleLeastVisitedSort
    - ArticleMostRatedSort
    - ArticleLeastRatedSort
    - ArticleMostReviewedSort
    - ArticleLeastReviewedSort
...
  utils.Result:
    properties:
      data: {}
      message:
        type: string
      statusCode:
        type: integer
      success:
        type: boolean
      systemMessage:
        type: string
    type: object
...
host: localhost:8585
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is a sample server.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: My API
  version: "1.0"
paths:
...
  /article:
    delete:
      consumes:
      - application/json
      description: Deletes an existing article by its ID
      parameters:
      - description: Article ID to delete
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/article.DeleteArticleCommand'
      produces:
      - application/json
      responses:
        "200":
          description: Deleted article confirmation
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "404":
          description: Article not found
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: Delete an article
      tags:
      - article
    get:
      consumes:
      - application/json
      description: Retrieves a specific article by its ID
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Article details
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "404":
          description: Article not found
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: Get article by ID
      tags:
      - article
    post:
      consumes:
      - application/json
      description: Creates a new article with the provided information
      parameters:
      - description: Article information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/article.CreateArticleCommand'
      produces:
      - application/json
      responses:
        "201":
          description: Created article
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: Create a new article
      tags:
      - article
    put:
      consumes:
      - application/json
      description: Updates an existing article with the provided information
      parameters:
      - description: Updated article information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/article.UpdateArticleCommand'
      produces:
      - application/json
      responses:
        "200":
          description: Updated article
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "404":
          description: Article not found
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: Update an article
      tags:
      - article
...
  /article/admin/all:
    get:
      consumes:
      - application/json
      description: Admin endpoint to retrieve all articles with additional information
      parameters:
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - in: query
        name: search
        type: string
      - enum:
        - title
        - description
        in: query
        name: searchBy
        type: string
      - enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - enum:
        - title
        - description
        in: query
        name: sortBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of all articles
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: 'Admin: Get all articles'
      tags:
      - article
  /article/all:
    get:
      consumes:
      - application/json
      description: Retrieves all articles with optional filtering
      parameters:
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - in: query
        name: search
        type: string
      - enum:
        - title
        - description
        in: query
        name: searchBy
        type: string
      - in: query
        name: siteId
        required: true
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - enum:
        - title
        - description
        in: query
        name: sortBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of articles
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: Get all articles
      tags:
      - article
  /article/filters-sort:
    post:
      consumes:
      - application/json
      description: Retrieves articles based on specified filters and sorting criteria
      parameters:
      - description: Filter and sort parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/article.GetByFiltersSortArticleQuery'
      produces:
      - application/json
      responses:
        "200":
          description: Filtered and sorted articles
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: Get articles by filters and sorting
      tags:
      - article
...
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"